Move TaskFragment.kt and AbstractTaskViewModel.kt under tasks/
