+++ ground/src/main/java/com/google/android/ground/ui/datacollection/tasks/multiplechoice/MultipleChoiceTaskViewModel.kt
@@ -19,7 +19,7 @@ import android.content.res.Resources
 import com.google.android.ground.model.submission.MultipleChoiceTaskData
 import com.google.android.ground.model.submission.MultipleChoiceTaskData.Companion.fromList
 import com.google.android.ground.model.task.Option
-import com.google.android.ground.ui.datacollection.AbstractTaskViewModel
+import com.google.android.ground.ui.datacollection.tasks.AbstractTaskViewModel
 import java8.util.Optional
 import javax.inject.Inject
