+++ ground/src/main/java/com/google/android/ground/ui/datacollection/tasks/number/NumberTaskViewModel.kt
@@ -17,7 +17,7 @@ package com.google.android.ground.ui.datacollection.tasks.number
 
 import android.content.res.Resources
 import com.google.android.ground.model.submission.NumberTaskData.Companion.fromNumber
-import com.google.android.ground.ui.datacollection.AbstractTaskViewModel
+import com.google.android.ground.ui.datacollection.tasks.AbstractTaskViewModel
 import javax.inject.Inject
 
 class NumberTaskViewModel @Inject constructor(resources: Resources) :
