+++ ground/src/main/res/layout/data_collection_header.xml
@@ -23,7 +23,7 @@
     <import type="android.view.View" />
     <variable
       name="viewModel"
-      type="com.google.android.ground.ui.datacollection.AbstractTaskViewModel" />
+      type="com.google.android.ground.ui.datacollection.tasks.AbstractTaskViewModel" />
   </data>
 
   <androidx.constraintlayout.widget.ConstraintLayout
